+++ server/CalcServer/Models/CalculationCommand.cs
@@ -6,12 +6,12 @@
     public class CalculationCommand
     {
         /// <summary> First of the values to be summated. </summary>
-        public float LeftOperand { get; set; }
+        public double LeftOperand { get; set; }
 
         /// <summary> Operation to perform. </summary>
         public string Operation { get; set; }
 
         /// <summary> Second of the values to be summated. </summary>
-        public float RightOperand { get; set; }
+        public double RightOperand { get; set; }
     }
 }
