+++ server/CalcServer/Services/ICalculator.cs
@@ -9,7 +9,7 @@ namespace CalcServer.Services
         /// <summary> Gets the solution for a given calculation command. </summary>
         /// <param name="command"> Command to execute. </param>
         /// <returns> The result of execution. </returns>
-        float Solve(CalculationCommand command);
+        double Solve(CalculationCommand command);
 
         /// <summary> Gets a list of supported operations. </summary>
         /// <returns> Operations supported by this calculator. </returns>
